app/backend/app.py
import os
from flask import Flask, jsonify, request
from google.cloud import healthcare
from google.cloud import aiplatform

app = Flask(__name__)

# The "Adaptive Synergy Optimization" (ASO) Logic from Doc 2
@app.route('/api/synergy/score', methods=['POST'])
def calculate_recovery_capital():
    data = request.json
    # C_i(t): Clinical indicators
    # P_i(t): Passive income/financial stability indicators
    
    clinical_score = data.get('clinical_adherence', 0)
    financial_score = data.get('passive_income_generated', 0)
    
    # Mathematical uplift formula
    synergy_score = (clinical_score * 0.6) + (financial_score * 0.4)
    
    return jsonify({
        "recovery_capital": synergy_score,
        "status": "OPTIMIZED" if synergy_score > 80 else "NEEDS_SUPPORT"
    })

# Digital Twin Generator
@app.route('/api/twin/generate', methods=['POST'])
def generate_twin():
    user_id = request.json['user_id']
    # 1. Fetch PHI from Healthcare API (Securely)
    # 2. Feed into Vertex AI Model
    # 3. Generate .usd/.glb asset
    # 4. Upload to signed URL in Cloud Storage
    
    return jsonify({"twin_asset_url": f"https://storage.googleapis.com/ihep-twins/{user_id}.glb"})

if __name__ == "__main__":
    app.run(host='0.0.0.0', port=int(os.environ.get('PORT', 8080)))