app/backend/requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.1
Flask-Limiter==3.5.0
gunicorn==22.0.1
gevent==23.9.1

# Google Cloud
google-cloud-healthcare==4.3.0
google-cloud-storage==2.14.0
google-cloud-sql==1.2.0
google-cloud-secret-manager==2.17.0
google-cloud-kms==2.20.0
google-cloud-monitoring==2.18.0
google-cloud-logging==3.9.0
google-cloud-trace==1.12.0
google-auth==2.27.0
google-auth-httplib2==0.2.0

# AI/ML
google-cloud-aiplatform==1.42.0
numpy==1.26.3
scipy==1.12.0
scikit-learn==1.4.0
pandas==2.2.0
torch==2.6.0
transformers==4.45.2

# Database
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.1

# Security
cryptography==42.0.6
PyJWT==2.8.0
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0.post1
bcrypt==4.1.2

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
marshmallow==3.20.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-flask==1.3.0
faker==22.5.0
hypothesis==6.98.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pylint==3.0.3
isort==5.13.2
bandit==1.7.6

# Utilities
python-dotenv==1.0.1
click==8.1.7
requests==2.31.0
httpx==0.26.0
tenacity==8.2.3
aiohttp==3.9.5
celery==5.3.6
kombu==5.3.5

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-flask==0.43b0
opentelemetry-exporter-gcp-trace==1.6.0

# Date/Time
python-dateutil==2.8.2
pytz==2024.1

# Math/Scientific
sympy==1.12
statsmodels==0.14.1

# Development
ipython==8.20.0
ipdb==0.13.13
